issue#10 Add initRoutes
