+++ app.js
@@ -3,15 +3,14 @@ import express from 'express';
 import path from 'path';
 import cookieParser from 'cookie-parser';
 import logger from 'morgan';
+import initRoutes from './src/routes';
 
-import indexRouter from './src/routes/index';
-import userApiRouter from './src/routes/userApi';
-// var usersRouter = require('./routes/users');
+
 
 var app = express();
 
 // view engine setup
-app.set('views', path.join(__dirname, 'views'));
+app.set('views', path.join(__dirname, './src/views'));
 app.set('view engine', 'ejs');
 
 app.use(logger('dev'));
@@ -20,10 +19,7 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/', indexRouter);
-app.use('/api/users', userApiRouter);
-// app.use('/users', usersRouter); // todo Add userList page (ejs)
-
+initRoutes(app);
 
 // Connect DB
 const db = require("./src/models");
